chore(node): remove NodeJS.Global
