+++ types/amqp/amqp-tests.ts
@@ -61,10 +61,4 @@ async function start() {
   }
 }
 
-async function wait(ms: number) {
-  return new Promise(resolve => {
-    setTimeout(resolve, ms);
-  });
-}
-
 start();
