+++ types/node/test/globals.ts
@@ -3,10 +3,26 @@
 //////////////////////////////////////////////////////////////////////////
 
 {
-    const isGlobal: NodeJS.Global = global;
     const isGlobalThis: typeof globalThis = global;
 
     const accessibleToGlobalThisMembers: true = global.RANDOM_GLOBAL_VARIABLE;
 }
 
 declare var RANDOM_GLOBAL_VARIABLE: true;
+
+// global aliases for compatibility
+{
+    const x: NodeModule = {} as any;
+    const y: NodeModule = {} as any;
+    x.children.push(y);
+    x.parent = require.main!;
+    require.main = y;
+    x.path; // $ExpectType string
+}
+
+// exposed gc
+{
+    if (gc) {
+        gc();
+    }
+}
