+++ types/node/timers.d.ts
@@ -1,5 +1,6 @@
 declare module 'timers' {
     import { Abortable } from 'events';
+    import { setTimeout as setTimeoutPromise, setImmediate as setImmediatePromise, setInterval as setIntervalPromise } from 'timers/promises';
 
     interface TimerOptions extends Abortable {
         /**
@@ -10,22 +11,82 @@ declare module 'timers' {
         ref?: boolean;
     }
 
-    function setTimeout(callback: (...args: any[]) => void, ms?: number, ...args: any[]): NodeJS.Timeout;
-    namespace setTimeout {
-        function __promisify__(ms: number): Promise<void>;
-        function __promisify__<T>(ms: number, value: T, options?: TimerOptions): Promise<T>;
+    let setTimeout: typeof global.setTimeout;
+    let clearTimeout: typeof global.clearTimeout;
+    let setInterval: typeof global.setInterval;
+    let clearInterval: typeof global.clearInterval;
+    let setImmediate: typeof global.setImmediate;
+    let clearImmediate: typeof global.clearImmediate;
+
+    global {
+        namespace NodeJS {
+            // compatibility with older typings
+            interface Timer extends RefCounted {
+                hasRef(): boolean;
+                refresh(): this;
+                [Symbol.toPrimitive](): number;
+            }
+
+            interface Immediate extends RefCounted {
+                hasRef(): boolean;
+                _onImmediate: Function; // to distinguish it from the Timeout class
+            }
+
+            interface Timeout extends Timer {
+                hasRef(): boolean;
+                refresh(): this;
+                [Symbol.toPrimitive](): number;
+            }
+        }
+
+        function setTimeout<TArgs extends any[]>(callback: (...args: TArgs) => void, ms?: number, ...args: TArgs): NodeJS.Timeout;
+        namespace setTimeout {
+            const __promisify__: typeof setTimeoutPromise;
+        }
+        function clearTimeout(timeoutId: NodeJS.Timeout): void;
+
+        function setInterval<TArgs extends any[]>(callback: (...args: TArgs) => void, ms?: number, ...args: TArgs): NodeJS.Timer;
+        namespace setInterval {
+            const __promisify__: typeof setIntervalPromise;
+        }
+        function clearInterval(intervalId: NodeJS.Timeout): void;
+
+        function setImmediate<TArgs extends any[]>(callback: (...args: TArgs) => void, ...args: TArgs): NodeJS.Immediate;
+        namespace setImmediate {
+            const __promisify__: typeof setImmediatePromise;
+        }
+        function clearImmediate(immediateId: NodeJS.Immediate): void;
+
+        function queueMicrotask(callback: () => void): void;
     }
-    function clearTimeout(timeoutId: NodeJS.Timeout): void;
-    function setInterval(callback: (...args: any[]) => void, ms?: number, ...args: any[]): NodeJS.Timeout;
-    function clearInterval(intervalId: NodeJS.Timeout): void;
-    function setImmediate(callback: (...args: any[]) => void, ...args: any[]): NodeJS.Immediate;
-    namespace setImmediate {
-        function __promisify__(): Promise<void>;
-        function __promisify__<T>(value: T, options?: TimerOptions): Promise<T>;
-    }
-    function clearImmediate(immediateId: NodeJS.Immediate): void;
 }
 
 declare module 'node:timers' {
     export * from 'timers';
+}
+
+declare module 'timers/promises' {
+    import { TimerOptions } from 'timers';
+
+    /**
+     * Returns a promise that resolves after the specified delay in milliseconds.
+     * @param delay defaults to 1
+     */
+    function setTimeout<T = void>(delay?: number, value?: T, options?: TimerOptions): Promise<T>;
+
+    /**
+     * Returns a promise that resolves in the next tick.
+     */
+    function setImmediate<T = void>(value?: T, options?: TimerOptions): Promise<T>;
+
+    /**
+     *
+     * Returns an async iterator that generates values in an interval of delay ms.
+     * @param delay defaults to 1
+     */
+    function setInterval<T = void>(delay?: number, value?: T, options?: TimerOptions): AsyncIterable<T>;
+}
+
+declare module 'node:timers/promises' {
+    export * from 'timers/promises';
 }
