+++ types/tile-reduce/tile-reduce-tests.ts
@@ -1,5 +1,5 @@
 import tilereduce = require('tile-reduce');
-import { BBox, Tile, Source, Types } from 'tile-reduce'
+import { BBox, Tile, Source, Types, } from 'tile-reduce'
 
 global.mapOptions
